// src/types/btn.type.ts
export type cornerType = 'square' | 'rounded' | 'pill'
export type btnType = 'submit' | 'reset' | 'button'
export type sizeBtnType = 'normal' | 'small' | 'large'
export type BgColorType =
  | 'transparent'
  | 'white'
  | 'black'
  | 'pink'
  | 'red'
  | 'blue'
  | 'green'
  | 'yellow'
  | 'gray'
  | 'purple'
export type ColorType =
  | 'transparent'
  | 'white'
  | 'black'
  | 'pink'
  | 'red'
  | 'blue'
  | 'green'
  | 'yellow'
  | 'gray'
  | 'purple'